pyproject.toml
[tool.poetry]
name = "metal-ads"
version = "0.1.0"
description = "Tools for performing metal-substrate adsorption calculations"
authors = ["Dylan Morgan <dylan.morgan@warwick.ac.uk>"]
license = "GPL"
readme = "README.md"
packages = [{include = "metal_ads"}]

[tool.poetry.dependencies]
python = "^3.8"
click = "^8.1.3"
PyYAML = "^6.0"

[tool.poetry.group.dev.dependencies]
ptvsd = "^4.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
